pi2/requirements.txt
absl-py==2.4.0
astunparse==1.6.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
colorzero==2.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.62.0
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
gpiozero==2.0.1
grpcio==1.78.0
h5py==3.14.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.13.2
kiwisolver==1.5.0
lgpio==0.2.2.0
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.4.3
onnx==1.20.1
onnxruntime==1.24.3
onnxslim==0.1.87
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
polars==1.39.0
polars-runtime-32==1.39.0
protobuf==7.34.0
psutil==7.2.2
PyAudio==0.2.14
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
sympy==1.14.0
tensorflow==2.21.0
tensorflow-hub==0.16.1
termcolor==3.3.0
tf2onnx==1.17.0
tf_keras==2.21.0
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
